correção nos modos de endereçamento

+++ include/Cpu.hpp
@@ -39,17 +39,17 @@ public:
   void reset();
 
   // Modos de endereçamento
-  uint8_t immediate();
-  uint8_t zeropage();
-  uint8_t zeropageX();
-  uint8_t absolute();
-  uint8_t absoluteX();
-  uint8_t absoluteY();
-  uint8_t indirectX();
-  uint8_t indirectY();
-  uint8_t accumulator();
+  uint16_t immediate();
+  uint16_t zeropage();
+  uint16_t zeropageX();
+  uint16_t absolute();
+  uint16_t absoluteX();
+  uint16_t absoluteY();
+  uint16_t indirectX();
+  uint16_t indirectY();
+  uint16_t accumulator();
 
-  // Funções de verificação de ativação de Flags após a operação
+  // Funções de verificação de ativação de Flags à partir de valores
   void flagActivationN(uint8_t value);
   void flagActivationV(uint8_t value_orig, uint8_t value_new);
   void flagActivationB();
@@ -60,13 +60,13 @@ public:
 
   // Implementações das instruções
   // ADC (ADd with Carry)
-  void ADC(uint8_t (Cpu::*AddressingMode)());
+  void ADC(uint16_t (Cpu::*AddressingMode)());
 
   // AND (bitwise AND with accumulator)
-  void AND(uint8_t (Cpu::*AddressingMode)());
+  void AND(uint16_t (Cpu::*AddressingMode)());
 
   // ASL (Arithmetic Shift Left)
-  void ASL(uint8_t (Cpu::*AddressingMode)());
+  void ASL(uint16_t (Cpu::*AddressingMode)());
 
   // BIT (test BITs)
   // Branch Instructions
@@ -78,13 +78,13 @@ public:
   // EOR (bitwise Exclusive OR)
 
   // Flag (Processor Status) Instructions
-  void CLC(uint8_t (Cpu::*AddressingMode)());
-  void SEC(uint8_t (Cpu::*AddressingMode)());
-  void CLI(uint8_t (Cpu::*AddressingMode)());
-  void SEI(uint8_t (Cpu::*AddressingMode)());
-  void CLV(uint8_t (Cpu::*AddressingMode)());
-  void CLD(uint8_t (Cpu::*AddressingMode)());
-  void SED(uint8_t (Cpu::*AddressingMode)());
+  void CLC(uint16_t (Cpu::*AddressingMode)());
+  void SEC(uint16_t (Cpu::*AddressingMode)());
+  void CLI(uint16_t (Cpu::*AddressingMode)());
+  void SEI(uint16_t (Cpu::*AddressingMode)());
+  void CLV(uint16_t (Cpu::*AddressingMode)());
+  void CLD(uint16_t (Cpu::*AddressingMode)());
+  void SED(uint16_t (Cpu::*AddressingMode)());
 
   // INC (INCrement memory)
   // JMP (JuMP)
